[add] 물품 등록 및 경매 자동 등록 internal api 추가

+++ rest_server/controllers/commonapi/commonapi_auc/api_auc_auction.go
@@ -40,10 +40,15 @@ func PostAucAuctionUpdate(auction *context_auc.AucAuctionUpdate, ctx *context.IP
 		log.Error("UpdateAucAuction :", err)
 		resp.SetReturn(resultcode.Result_DBError)
 	} else {
-		auction.Id = id
-		resp.Value = auction
-		// redis 삭제
-		model.GetDB().DeleteAuctionCache(auction.Id)
+		if id == 0 {
+			resp.SetReturn(resultcode.Result_DBNotExistAuction)
+		} else {
+			auction.Id = id
+			resp.Value = auction
+			// redis 삭제
+			model.GetDB().DeleteAuctionCache(auction.Id)
+		}
+
 	}
 	return ctx.EchoContext.JSON(http.StatusOK, resp)
 }
@@ -223,17 +228,29 @@ func PostAucAuctionFinish(auctionFinish *context_auc.AuctionFinish, ctx *context
 	resp.Success()
 
 	// 1. 경매 테이블 종료 업데이트
-	if affected, err := model.GetDB().UpdateAucAuctionFinish(auctionFinish.Id, context_auc.Auction_auc_state_finish); err != nil {
+	if affected, err := model.GetDB().UpdateAucAuctionAucState(auctionFinish.Id, context_auc.Auction_auc_state_finish); err != nil {
 		resp.SetReturn(resultcode.Result_DBError)
 	} else if err == nil && affected == 0 {
 		resp.SetReturn(resultcode.Result_DBNotExistAuction)
 	} else {
-		// 2. 입찰자 리스트 낙찰 업데이트
+		// 2. 기존 최고 입찰자 정보 가져오기
+		bid, err := model.GetDB().GetAucBidBestAttendee(auctionFinish.Id)
+		if err != nil {
+			log.Error("PostAucBidSubmit :", err)
+			resp.SetReturn(resultcode.Result_DBError)
+		} else {
+			// 3. 입찰자 리스트 낙찰 업데이트
+			if _, err := model.GetDB().UpdateAucBidWinnerState(bid, context_auc.Bid_state_success); err != nil {
+				log.Error("PostAucBidSubmit :", err)
+				resp.SetReturn(resultcode.Result_DBError)
+			} else {
+				// 4. 보증금 반환 리스트 반환
 
-		// 3. 보증금 반환 리스트 반환
+				// 5. redis 삭제
+				model.GetDB().DeleteAuctionCache(auctionFinish.Id)
+			}
+		}
 
-		// redis 삭제
-		model.GetDB().DeleteAuctionCache(auctionFinish.Id)
 	}
 
 	return ctx.EchoContext.JSON(http.StatusOK, resp)
